expensify/requirements/base-pip-requirements.txt
asgiref==3.2.7
Django==3.0.8
django-filter==2.2.0
django-extensions==2.2.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
pytz==2019.3
sqlparse==0.3.1
celery==4.3.0
Fabric3==1.14.post1
psycopg2-binary
query-string==2019.4.13
requests==2.23.0
redis==3.4.1
django-cors-headers==3.4.0